Revert "Dotnet folder restructure and initial UI layout"
