+++ dotnet/App.axaml.cs
@@ -3,12 +3,8 @@ using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Data.Core;
 using Avalonia.Data.Core.Plugins;
 using Avalonia.Markup.Xaml;
-using dotnet.Common.Data;
-using dotnet.Common.Presentation;
-using dotnet.Features.Dashboard.Presentation;
-using Microsoft.EntityFrameworkCore;
-using System;
-
+using dotnet.ViewModels;
+using dotnet.Views;
 
 namespace dotnet;
 
@@ -16,26 +12,22 @@ public partial class App : Application
 {
     public override void Initialize()
     {
-        InitializeDatabase();
         AvaloniaXamlLoader.Load(this);
     }
 
-    private void InitializeDatabase()
-    {
-        using (var dbContext = new AppDbContext())
-        {
-            dbContext.Database.Migrate(); // Apply any migrations
-                                          // Log for verification
-            Console.WriteLine("Database initialized and migrations applied.");
-        }
-    }
     public override void OnFrameworkInitializationCompleted()
     {
-            if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
-                desktop.MainWindow = new MainWindow();
-            else if (ApplicationLifetime is ISingleViewApplicationLifetime singleView)
-                singleView.MainView = new DashboardSingleView();
+        if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
+        {
+            // Line below is needed to remove Avalonia data validation.
+            // Without this line you will get duplicate validations from both Avalonia and CT
+            BindingPlugins.DataValidators.RemoveAt(0);
+            desktop.MainWindow = new MainWindow
+            {
+                DataContext = new MainWindowViewModel(),
+            };
+        }
 
-            base.OnFrameworkInitializationCompleted();
+        base.OnFrameworkInitializationCompleted();
     }
 }
