--- dotnet/Common/ErrorHandling/IEither.cs
@@ -1,7 +1,0 @@
-namespace dotnet.Common.ErrorHandling
-{
-    public interface IEither
-    {
-
-    }
-}
