--- dotnet/Common/ErrorHandling/Response.cs
@@ -1,15 +1,0 @@
-namespace dotnet.Common.ErrorHandling
-{
-    public class Response<T>
-    {
-        public bool IsSuccess { get; }
-
-        public T Data { get;}
-
-        public Response(T data)
-        {
-            IsSuccess = true;
-            Data = data;
-        }
-    }
-}
