--- dotnet/Common/Presentation/MainWindow.axaml
@@ -1,12 +1,0 @@
-<Window xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:dashboardViews="clr-namespace:dotnet.Features.Dashboard.Presentation"
-        mc:Ignorable="d"
-        d:DesignWidth="582"
-        d:DesignHeight="372"
-        x:Class="dotnet.Common.Presentation.MainWindow"
-        Title="dotnet">
-    <dashboardViews:DashboardView />
-</Window>
