--- dotnet/Common/Presentation/MainWindowViewModel.cs
@@ -1,8 +1,0 @@
-namespace dotnet.Common.Presentation;
-
-public partial class MainWindowViewModel : ViewModelBase
-{
-#pragma warning disable CA1822 // Mark members as static
-    public string Greeting => "Welcome to Avalonia!";
-#pragma warning restore CA1822 // Mark members as static
-}
