--- dotnet/Common/Presentation/ViewModelBase.cs
@@ -1,7 +1,0 @@
-using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace dotnet.Common.Presentation;
-
-public class ViewModelBase : ObservableObject
-{
-}
