--- dotnet/Features/Dashboard/Presentation/DashboardSingleView.axaml
@@ -1,11 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:dashboardViews="clr-namespace:dotnet.Features.Dashboard.Presentation"
-             mc:Ignorable="d"
-             d:DesignWidth="800"
-             d:DesignHeight="450"
-             x:Class="dotnet.Features.Dashboard.Presentation.DashboardSingleView">
-    <dashboardViews:DashboardSingleView />
-</UserControl>
