--- dotnet/Features/Dashboard/Presentation/DashboardView.axaml.cs
@@ -1,15 +1,0 @@
-using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-using SkiaSharp;
-using Svg.Skia;
-
-namespace dotnet.Features.Dashboard.Presentation;
-
-public partial class DashboardView : UserControl
-{
-    public DashboardView()
-    {
-        InitializeComponent();
-    }
-}
