--- dotnet/Features/DeviceConnectivity/I2C/I2C.Tests/GlobalUsings.cs
@@ -1,1 +1,0 @@
-global using Xunit;
