--- dotnet/Features/DeviceConnectivity/Mqtt/IMqttClient.cs
@@ -1,7 +1,0 @@
-namespace dotnet.Features.DeviceConnectivity.Mqtt
-{
-    public interface IMqttClient
-    {
-        
-    }
-}
