--- dotnet/Features/DeviceConnectivity/Mqtt/Implementations/MqttClient.cs
@@ -1,7 +1,0 @@
-namespace dotnet.Features.DeviceConnectivity.Mqtt.Implementations
-{
-    public class MqttClient : IMqttClient
-    {
-        
-    }
-}
