--- dotnet/Features/SensorManagement/TemperatureSensor/Domain/Models/TemperatureSensor.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace dotnet.Features.SensorManagement.TemperatureSensor.Domain.Models
-{
-    public class TemperatureSensor
-    {
-        // Id for the temperature sensor
-        public int Id { get; set; }
-
-        /* Location of the sensor (ex. water tank, air, can be also TYPE of temperature sensor.)
-        public string Location { get; set; }*/
-
-        // Current temperature reading 
-        public double CurrentTemperature { get; set; }
-
-        // Timestamp of the last reading
-        public DateTime LastReadingTime { get; set; }
-
-        //  update the temperature reading mnethod
-        public void UpdateTemperature(double newTemperature)
-        {
-            CurrentTemperature = newTemperature;
-            LastReadingTime = DateTime.Now; 
-        }
-
-        
-    }
-}
