+++ dotnet/ViewModels/MainWindowViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace dotnet.ViewModels;
+
+public partial class MainWindowViewModel : ViewModelBase
+{
+#pragma warning disable CA1822 // Mark members as static
+    public string Greeting => "Welcome to Avalonia!";
+#pragma warning restore CA1822 // Mark members as static
+}
