+++ dotnet/ViewModels/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace dotnet.ViewModels;
+
+public class ViewModelBase : ObservableObject
+{
+}
