--- dotnet/dotnet.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35222.181
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotnet", "dotnet.csproj", "{337E8187-F7F2-44A6-BC3E-59C5E2893D4B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{337E8187-F7F2-44A6-BC3E-59C5E2893D4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{337E8187-F7F2-44A6-BC3E-59C5E2893D4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{337E8187-F7F2-44A6-BC3E-59C5E2893D4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{337E8187-F7F2-44A6-BC3E-59C5E2893D4B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B3972399-E9D6-4BBD-89F6-8DD4C1AAEB5E}
-	EndGlobalSection
-EndGlobal
