+++ esp32_devices/temperature_sensor/main/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "temperature_sensor.c"
+                    INCLUDE_DIRS ".")
