+++ esp32_devices/temperature_sensor/main/temperature_sensor.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+void app_main(void)
+{
+
+}
