+++ esp32_devices/water_pump/main/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "water_pump.c"
+                    INCLUDE_DIRS ".")
