+++ esp32_devices/water_pump/main/water_pump.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+void app_main(void)
+{
+
+}
